requirements.txt
streamlit
opencv-python-headless
numpy
pandas
librosa
torch
Pillow
scikit-learn
tensorflow
transformers
spotipy
